# webapp/openapi.yaml
openapi: 3.0.0
info:
  title: Signature Verification API
  version: 1.0.0
servers:
  - url: http://localhost:5217
paths:
  /signature/detect:
    post:
      operationId: detectSignature
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      parameters:
        - name: includeImages
          in: query
          schema:
            type: boolean
          required: false
        - name: enableYoloV8
          in: query
          schema:
            type: boolean
          required: false
        - name: enableDetr
          in: query
          schema:
            type: boolean
          required: false
        - name: strategy
          in: query
          schema:
            type: string
          required: false
        - name: yoloConfidenceThreshold
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: yoloNmsIoU
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: detrConfidenceThreshold
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: fallbackFp
          in: query
          schema:
            type: integer
          required: false
        - name: fallbackFn
          in: query
          schema:
            type: integer
          required: false
        - name: eceDetr
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: eceYolo
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: ensembleThreshold
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: enableShapeRoiV2
          in: query
          schema:
            type: boolean
          required: false
        - name: shapeMinAspect
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: shapeMaxAspect
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: lowConfidence
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: highConfidence
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: cropMarginPerc
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: roiConfirmIoU
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: uncertainQuantile
          in: query
          schema:
            type: number
            format: float
          required: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectResponseDto'
  /signature/verify:
    post:
      operationId: verifySignature
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                reference:
                  type: string
                  format: binary
                candidate:
                  type: string
                  format: binary
      parameters:
        - name: detection
          in: query
          schema:
            type: boolean
          required: false
        - name: threshold
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: preprocessed
          in: query
          schema:
            type: boolean
          required: false
        - name: enableYoloV8
          in: query
          schema:
            type: boolean
          required: false
        - name: enableDetr
          in: query
          schema:
            type: boolean
          required: false
        - name: strategy
          in: query
          schema:
            type: string
          required: false
        - name: yoloConfidenceThreshold
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: yoloNmsIoU
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: detrConfidenceThreshold
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: fallbackFp
          in: query
          schema:
            type: integer
          required: false
        - name: fallbackFn
          in: query
          schema:
            type: integer
          required: false
        - name: eceDetr
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: eceYolo
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: ensembleThreshold
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: enableShapeRoiV2
          in: query
          schema:
            type: boolean
          required: false
        - name: shapeMinAspect
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: shapeMaxAspect
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: lowConfidence
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: highConfidence
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: cropMarginPerc
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: roiConfirmIoU
          in: query
          schema:
            type: number
            format: float
          required: false
        - name: uncertainQuantile
          in: query
          schema:
            type: number
            format: float
          required: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponseDto'
components:
  schemas:
    BoundingBoxDto:
      type: object
      properties:
        x1:
          type: number
        y1:
          type: number
        x2:
          type: number
        y2:
          type: number
    SignatureDto:
      type: object
      properties:
        confidence:
          type: number
        boundingBox:
          $ref: '#/components/schemas/BoundingBoxDto'
        imageData:
          type: string
          format: byte
          nullable: true
    DetectResponseDto:
      type: object
      properties:
        signatures:
          type: array
          items:
            $ref: '#/components/schemas/SignatureDto'
    VerifyResponseDto:
      type: object
      properties:
        forged:
          type: boolean
        similarity:
          type: number
        referenceImage:
          type: string
          format: byte
          nullable: true
        candidateImage:
          type: string
          format: byte
          nullable: true